pyproject.toml
[project]
name = "llm-demo"
version = "0.1.0"
description = "A Flask chatbot UI with streaming OpenAI responses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0"
]
